Refactor protos, add deserialisation for lots of classes

+++ server/core/src/main/scala/org/oliverlittle/clusterprocess/connector/PingLatency.scala
@@ -51,6 +51,6 @@ object PingLatency {
             case e : IOException => println(e)
         }
 
-        return Some(stop - start)
+        return Some(stop.toDouble - start.toDouble)
     }
 }
